Solution to Check Less Than

+++ Check Less Than@@67882a731b4392d8fc1577d7/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    float x,y;
-    scanf("%f %f",&x,&y);
+    int x,y;
+    scanf("%d %d",&x,&y);
     if(x>y){
         printf("False\n");
     }
